+++ src/charade/components/charade-play/charade-play.tsx
@@ -1,15 +1,10 @@
 import * as React from 'react'
-import { useTheme } from 'styled-components/native'
-import { SafeAreaFlexContainer } from '@core/styled'
-import { ThemeModel } from '@core/styled/models'
-import { Header } from '@core/header/header'
-import { CharadePlayground } from '../charade-playground/charade-playground'
-import { useCharadeWords } from '../../hooks'
-import { useCharadeSpy } from 'charade/hooks/use-charade-spy.hook'
 import { leaveGameAlert } from '@core/alerts/leave-game-alert'
+import { Template } from '@core/template/template'
+import { CharadePlayground } from '../charade-playground/charade-playground'
+import { useCharadeWords, useCharadeSpy } from '../../hooks'
 
 export const CharadePlay = () => {
-  const theme = useTheme() as ThemeModel
   const { word, allWords } = useCharadeWords()
   const [end, setEnd] = React.useState(false)
 
@@ -18,9 +13,8 @@ export const CharadePlay = () => {
   const backButtonAlert = end ? undefined : leaveGameAlert
 
   return (
-    <SafeAreaFlexContainer backgroundColor={theme.backgroundPrimary}>
-      <Header type="charade" onTouchEnd={onTouchEnd} backButtonAlert={backButtonAlert} backButton />
+    <Template type="charade" onTouchEnd={onTouchEnd} backButtonAlert={backButtonAlert} backButton flex>
       <CharadePlayground {...{ word, allWords, setEnd }} />
-    </SafeAreaFlexContainer>
+    </Template>
   )
 }