+++ src/more/components/empty-options/empty-options.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { ActivityIndicator } from 'react-native'
-import { MarginView } from '@core/styled/margin-view.styled'
+import { MarginView } from '@core/styled'
 
 export const EmptyOptions = (
   <MarginView margins={[ 30, 0, 0, 0 ]}>